+++ tt/client_test.go
@@ -141,12 +141,12 @@ type Interceptor struct {
 }
 
 func (r *Interceptor) intercept(next mq.Handler) mq.Handler {
-	return func(p mq.Packet) error {
+	return func(ctx Context, p mq.Packet) error {
 		select {
 		case r.c <- p: // if anyone is interested
 		case <-time.After(10 * time.Millisecond):
 		}
-		return next(p)
+		return next(ctx, p)
 	}
 }